+++ stencil-workspace/src/components/modus-content-tree/modus-content-tree.e2e.ts
@@ -145,7 +145,7 @@ describe('modus-tree-view-item', () => {
     </modus-tree-view>`);
     const root = await page.find('modus-tree-view');
 
-    const rightChevron = await page.find('modus-tree-view-item >>> .icon-chevron-right-thick');
+    const rightChevron = await page.find('modus-tree-view-item >>> .rotate-right .icon-chevron-down-thick');
     expect(rightChevron).toBeTruthy();
 
     root.setProperty('expandedItems', ['1']);
@@ -184,7 +184,7 @@ describe('modus-tree-view-item', () => {
       </modus-tree-view-item>
     </modus-tree-view>`);
 
-    const rightChevron = await page.find('modus-tree-view-item[node-id=\'1\'] >>> .icon-chevron-right-thick');
+    const rightChevron = await page.find('modus-tree-view-item[node-id=\'1\'] >>> .rotate-right .icon-chevron-down-thick');
     expect(rightChevron).toBeTruthy();
 
     await rightChevron.click();
@@ -393,7 +393,7 @@ describe('modus-tree-view-item', () => {
     </modus-tree-view>`);
 
     const itemExpandToggle = await page.spyOnEvent('itemExpandToggle');
-    const rightChevron = await page.find('modus-tree-view-item[node-id=\'1\'] >>> .icon-chevron-right-thick');
+    const rightChevron = await page.find('modus-tree-view-item[node-id=\'1\'] >>> .rotate-right .icon-chevron-down-thick');
     expect(rightChevron).toBeTruthy();
 
     await rightChevron.click();